Refactor controller and keymap sections
